Split DMA module into files by component
