+++ rp2040-hal/src/dma/mod.rs
@@ -24,10 +24,20 @@
 
 use crate::resets::SubsystemReset;
 use core::marker::PhantomData;
-use core::mem;
-use core::sync::atomic::{compiler_fence, Ordering};
 use embedded_dma::{ReadBuffer, WriteBuffer};
 use rp2040_pac::DMA;
+
+// Bring in our submodules
+mod bidirectional;
+mod double_buffering;
+mod ring_transfer;
+mod single_buffering;
+mod single_channel;
+
+// Export these types for easier use by external code
+pub use crate::dma::bidirectional::BidirectionalConfig;
+pub use crate::dma::double_buffering::DoubleBufferingConfig;
+pub use crate::dma::single_buffering::SingleBufferingConfig;
 
 /// DMA unit.
 pub trait DMAExt {
@@ -110,222 +120,6 @@ impl<CH: ChannelIndex> ChannelRegs for Channel<CH> {
 
     fn regs(&self) -> &rp2040_pac::dma::CH {
         unsafe { &*Self::ptr() }
-    }
-}
-
-/// Trait which implements low-level functionality for transfers using a single DMA channel.
-pub trait SingleChannel {
-    /// Returns the registers associated with this DMA channel.
-    ///
-    /// In the case of channel pairs, this returns the first channel.
-    fn ch(&self) -> &rp2040_pac::dma::CH;
-    /// Returns the index of the DMA channel.
-    fn id(&self) -> u8;
-
-    /// Enables the DMA_IRQ_0 signal for this channel.
-    fn listen_irq0(&mut self) {
-        const ATOMIC_SET_OFFSET: usize = 0x2000;
-        // Safety: We only use the atomic alias of the register.
-        unsafe {
-            (*DMA::ptr())
-                .inte0
-                .as_ptr()
-                .add(ATOMIC_SET_OFFSET / 4)
-                .write_volatile(1 << self.id());
-        }
-    }
-
-    /// Disables the DMA_IRQ_0 signal for this channel.
-    fn unlisten_irq0(&mut self) {
-        const ATOMIC_CLEAR_OFFSET: usize = 0x3000;
-        // Safety: We only use the atomic alias of the register.
-        unsafe {
-            (*DMA::ptr())
-                .inte1
-                .as_ptr()
-                .add(ATOMIC_CLEAR_OFFSET / 4)
-                .write_volatile(1 << self.id());
-        }
-    }
-
-    /// Check if an interrupt is pending for this channel and clear the corresponding pending bit
-    fn check_irq0(&mut self) -> bool {
-        // Safety: The following is race-free as we only ever clear the bit for this channel.
-        // Nobody else modifies that bit.
-        unsafe {
-            let status = (*DMA::ptr()).ints0.read().bits();
-            if (status & (1 << self.id())) != 0 {
-                // Clear the interrupt.
-                (*DMA::ptr()).ints0.write(|w| w.bits(1 << self.id()));
-                true
-            } else {
-                false
-            }
-        }
-    }
-
-    /// Enables the DMA_IRQ_0 signal for this channel.
-    fn listen_irq1(&mut self) {
-        const ATOMIC_SET_OFFSET: usize = 0x2000;
-        // Safety: We only use the atomic alias of the register.
-        unsafe {
-            (*DMA::ptr())
-                .inte1
-                .as_ptr()
-                .add(ATOMIC_SET_OFFSET / 4)
-                .write_volatile(1 << self.id());
-        }
-    }
-
-    /// Disables the DMA_IRQ_1 signal for this channel.
-    fn unlisten_irq1(&mut self) {
-        const ATOMIC_CLEAR_OFFSET: usize = 0x3000;
-        // Safety: We only use the atomic alias of the register.
-        unsafe {
-            (*DMA::ptr())
-                .inte1
-                .as_ptr()
-                .add(ATOMIC_CLEAR_OFFSET / 4)
-                .write_volatile(1 << self.id());
-        }
-    }
-
-    /// Check if an interrupt is pending for this channel and clear the corresponding pending bit
-    fn check_irq1(&mut self) -> bool {
-        // Safety: The following is race-free as we only ever clear the bit for this channel.
-        // Nobody else modifies that bit.
-        unsafe {
-            let status = (*DMA::ptr()).ints1.read().bits();
-            if (status & (1 << self.id())) != 0 {
-                // Clear the interrupt.
-                (*DMA::ptr()).ints1.write(|w| w.bits(1 << self.id()));
-                true
-            } else {
-                false
-            }
-        }
-    }
-}
-
-/// Trait which implements low-level functionality for transfers requiring two DMA channels.
-///
-/// Anything that requires more than a single buffer exactly once requires two channels to be
-/// combined.
-pub trait ChannelPair: SingleChannel {
-    /// Returns the registers associated with the second DMA channel associated with this channel
-    /// pair.
-    fn ch2(&self) -> &rp2040_pac::dma::CH;
-    /// Returns the index of the second DMA channel.
-    fn id2(&self) -> u8;
-}
-
-impl<CH: ChannelIndex> SingleChannel for Channel<CH> {
-    fn ch(&self) -> &rp2040_pac::dma::CH {
-        self.regs()
-    }
-
-    fn id(&self) -> u8 {
-        CH::id()
-    }
-}
-
-impl<CH1: ChannelIndex, CH2: ChannelIndex> SingleChannel for (Channel<CH1>, Channel<CH2>) {
-    fn ch(&self) -> &rp2040_pac::dma::CH {
-        self.0.regs()
-    }
-
-    fn id(&self) -> u8 {
-        CH1::id()
-    }
-}
-
-trait ChannelConfig {
-    fn config<WORD, FROM, TO>(
-        &mut self,
-        from: &FROM,
-        to: &mut TO,
-        pace: Pace,
-        chain_to: Option<u8>,
-        start: bool,
-    ) where
-        FROM: ReadTarget<ReceivedWord = WORD>,
-        TO: WriteTarget<TransmittedWord = WORD>;
-
-    fn set_chain_to_enabled<CH: SingleChannel>(&mut self, other: &mut CH);
-    fn start(&mut self);
-}
-
-impl<CH: SingleChannel> ChannelConfig for CH {
-    fn config<WORD, FROM, TO>(
-        &mut self,
-        from: &FROM,
-        to: &mut TO,
-        pace: Pace,
-        chain_to: Option<u8>,
-        start: bool,
-    ) where
-        FROM: ReadTarget<ReceivedWord = WORD>,
-        TO: WriteTarget<TransmittedWord = WORD>,
-    {
-        // Configure the DMA channel.
-        let (src, src_count) = from.rx_address_count();
-        let src_incr = from.rx_increment();
-        let (dest, dest_count) = to.tx_address_count();
-        let dest_incr = to.tx_increment();
-        const TREQ_UNPACED: u8 = 0x3f;
-        let treq = match pace {
-            Pace::PreferSource => FROM::rx_treq().or_else(TO::tx_treq).unwrap_or(TREQ_UNPACED),
-            Pace::PreferSink => TO::tx_treq().or_else(FROM::rx_treq).unwrap_or(TREQ_UNPACED),
-        };
-        let len = u32::min(src_count, dest_count);
-        self.ch().ch_al1_ctrl.write(|w| unsafe {
-            w.data_size().bits(mem::size_of::<WORD>() as u8 >> 1);
-            w.incr_read().bit(src_incr);
-            w.incr_write().bit(dest_incr);
-            w.treq_sel().bits(treq);
-            w.chain_to().bits(chain_to.unwrap_or_else(|| self.id()));
-            w.en().bit(true);
-            w
-        });
-        self.ch().ch_read_addr.write(|w| unsafe { w.bits(src) });
-        self.ch().ch_trans_count.write(|w| unsafe { w.bits(len) });
-        if start {
-            self.ch()
-                .ch_al2_write_addr_trig
-                .write(|w| unsafe { w.bits(dest) });
-        } else {
-            self.ch().ch_write_addr.write(|w| unsafe { w.bits(dest) });
-        }
-    }
-
-    fn set_chain_to_enabled<CH2: SingleChannel>(&mut self, other: &mut CH2) {
-        // We temporarily pause the channel when setting CHAIN_TO, to prevent any race condition
-        // that could occur, as we need to check afterwards whether the channel was successfully
-        // chained to this channel or whether this channel was already completed. If we did not
-        // pause this channel, we could get into a situation where both channels completed in quick
-        // succession, yet we did not notice, as the situation is not distinguishable from one
-        // where the second channel was not started at all.
-
-        self.ch().ch_al1_ctrl.modify(|_, w| unsafe {
-            w.chain_to().bits(other.id());
-            w.en().clear_bit();
-            w
-        });
-        if self.ch().ch_al1_ctrl.read().busy().bit_is_set() {
-            // This channel is still active, so just continue.
-            self.ch().ch_al1_ctrl.modify(|_, w| w.en().set_bit());
-        } else {
-            // This channel has already finished, so just start the other channel directly.
-            other.start();
-        }
-    }
-
-    fn start(&mut self) {
-        // Safety: The write does not interfere with any other writes, it only affects this
-        // channel.
-        unsafe { &*rp2040_pac::DMA::ptr() }
-            .multi_chan_trigger
-            .write(|w| unsafe { w.bits(1 << self.id()) });
     }
 }
 
@@ -421,37 +215,7 @@ impl<B: WriteBuffer> WriteTarget for B {
     }
 }
 
-/*SingleBuffered<CH, RX, TX> {
-    config(...) -> SingleBufferedConfig;
-    is_done() -> bool
-    wait() -> (CH1, RX, TX)
-}
-
-DoubleBuffered<(CH1, CH2), RX, TX, ()> {
-    config(...) -> SingleBufferedConfig;
-    is_done() -> bool
-    read_next(BUF) -> DoubleReadBuffer<CH, RX, TX, BUF>
-    write_next(BUF) -> DoubleWriteBuffer<CH, RX, TX, BUF>
-    wait() -> ((CH1, CH2), RX, TX)
-}
-DoubleBuffered<CH, RX, TX, ReadNext<RX2>> {
-    is_done() -> bool
-    wait() -> (DoubleBuffered<CH, RX2, TX>, RX)
-}
-DoubleBuffered<CH, RX, TX, WriteNext<TX2>> {
-    is_done() -> bool
-    wait() -> (DoubleBuffered<CH, RX2, TX>, RX)
-}
-
-Bidirectional<(CH1, CH2), FROM, BIDI, TO> {
-    config(...) -> BidirectionalConfig;
-    from_is_done() -> bool
-    to_is_done() -> bool
-    is_done() -> bool
-    from_wait() -> (CH1, FROM, SingleBuffered<CH2, BIDI, TO>)
-    to_wait() -> (CH2, SingleBuffered<CH2, FROM, BIDI>, TO)
-    wait() -> (CH1, CH2, FROM, BIDI, TO)
-}
+/*
 
 BidiDoubleBuffered<(CH1, CH2), FROM, BIDI, TO, (), ()> {
     config(...) -> BidirectionalConfig;
@@ -463,621 +227,6 @@ BidiDoubleBuffered<(CH1, CH2), FROM, BIDI, TO, (), ()> {
 
 Endless<(CH1, CH2), RX, TX> {
     stop() -> bool
-}*/
-
-/// Configuration for single-buffered DMA transfer
-pub struct SingleBufferingConfig<CH: SingleChannel, FROM: ReadTarget, TO: WriteTarget> {
-    ch: CH,
-    from: FROM,
-    to: TO,
-    pace: Pace,
-}
-
-impl<CH, FROM, TO, WORD> SingleBufferingConfig<CH, FROM, TO>
-where
-    CH: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Create a new configuration for single-buffered DMA transfer
-    pub fn new(ch: CH, from: FROM, to: TO) -> SingleBufferingConfig<CH, FROM, TO> {
-        SingleBufferingConfig {
-            ch,
-            from,
-            to,
-            pace: Pace::PreferSource,
-        }
-    }
-
-    /// Sets the (preferred) pace for the DMA transfers.
-    ///
-    /// Usually, the code will automatically configure the correct pace, but
-    /// peripheral-to-peripheral transfers require the user to manually select whether the source
-    /// or the sink shall be queried for the pace signal.
-    pub fn pace(&mut self, pace: Pace) {
-        self.pace = pace;
-    }
-
-    /// Start the DMA transfer
-    pub fn start(mut self) -> SingleBuffering<CH, FROM, TO> {
-        // TODO: Do we want to call any callbacks to configure source/sink?
-
-        // Make sure that memory contents reflect what the user intended.
-        // TODO: How much of the following is necessary?
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Configure the DMA channel and start it.
-        self.ch
-            .config(&self.from, &mut self.to, self.pace, None, true);
-
-        SingleBuffering {
-            ch: self.ch,
-            from: self.from,
-            to: self.to,
-        }
-    }
-}
-
-// TODO: Drop for most of these structs
-/// Instance of a single-buffered DMA transfer
-pub struct SingleBuffering<CH: SingleChannel, FROM: ReadTarget, TO: WriteTarget> {
-    ch: CH,
-    from: FROM,
-    to: TO,
-}
-
-impl<CH, FROM, TO, WORD> SingleBuffering<CH, FROM, TO>
-where
-    CH: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Check if an interrupt is pending for this channel and clear the corresponding pending bit
-    pub fn check_irq0(&mut self) -> bool {
-        self.ch.check_irq0()
-    }
-
-    /// Check if an interrupt is pending for this channel and clear the corresponding pending bit
-    pub fn check_irq1(&mut self) -> bool {
-        self.ch.check_irq1()
-    }
-
-    /// Check if the transfer has completed.
-    pub fn is_done(&self) -> bool {
-        !self.ch.ch().ch_ctrl_trig.read().busy().bit_is_set()
-    }
-
-    /// Block until the transfer is complete, returning the channel and targets
-    pub fn wait(self) -> (CH, FROM, TO) {
-        while !self.is_done() {}
-
-        // Make sure that memory contents reflect what the user intended.
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        (self.ch, self.from, self.to)
-    }
-}
-
-/// Configuration for double-buffered DMA transfer
-pub struct DoubleBufferingConfig<
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget,
-    TO: WriteTarget,
-> {
-    ch: (CH1, CH2),
-    from: FROM,
-    to: TO,
-    pace: Pace,
-}
-
-impl<CH1, CH2, FROM, TO, WORD> DoubleBufferingConfig<CH1, CH2, FROM, TO>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Create a new configuration for double-buffered DMA transfer
-    pub fn new(ch: (CH1, CH2), from: FROM, to: TO) -> DoubleBufferingConfig<CH1, CH2, FROM, TO> {
-        DoubleBufferingConfig {
-            ch,
-            from,
-            to,
-            pace: Pace::PreferSource,
-        }
-    }
-
-    /// Sets the (preferred) pace for the DMA transfers.
-    ///
-    /// Usually, the code will automatically configure the correct pace, but
-    /// peripheral-to-peripheral transfers require the user to manually select whether the source
-    /// or the sink shall be queried for the pace signal.
-    pub fn pace(&mut self, pace: Pace) {
-        self.pace = pace;
-    }
-
-    /// Start the DMA transfer
-    pub fn start(mut self) -> DoubleBuffering<CH1, CH2, FROM, TO, ()> {
-        // TODO: Do we want to call any callbacks to configure source/sink?
-
-        // Make sure that memory contents reflect what the user intended.
-        // TODO: How much of the following is necessary?
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Configure the DMA channel and start it.
-        self.ch
-            .0
-            .config(&self.from, &mut self.to, self.pace, None, true);
-
-        DoubleBuffering {
-            ch: self.ch,
-            from: self.from,
-            to: self.to,
-            pace: self.pace,
-            state: (),
-            second_ch: false,
-        }
-    }
-}
-
-/// State for a double-buffered read
-pub struct ReadNext<BUF: ReadTarget>(BUF);
-/// State for a double-buffered write
-pub struct WriteNext<BUF: WriteTarget>(BUF);
-
-/// Instance of a double-buffered DMA transfer
-pub struct DoubleBuffering<CH1, CH2, FROM, TO, STATE>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget,
-    TO: WriteTarget,
-{
-    ch: (CH1, CH2),
-    from: FROM,
-    to: TO,
-    pace: Pace,
-    state: STATE,
-    second_ch: bool,
-}
-
-impl<CH1, CH2, FROM, TO, WORD, STATE> DoubleBuffering<CH1, CH2, FROM, TO, STATE>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Check if the transfer is completed
-    pub fn is_done(&self) -> bool {
-        if self.second_ch {
-            !self.ch.1.ch().ch_ctrl_trig.read().busy().bit_is_set()
-        } else {
-            !self.ch.0.ch().ch_ctrl_trig.read().busy().bit_is_set()
-        }
-    }
-}
-
-impl<CH1, CH2, FROM, TO, WORD> DoubleBuffering<CH1, CH2, FROM, TO, ()>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD> + EndlessWriteTarget,
-{
-    /// Block until transfer completed
-    pub fn wait(self) -> (CH1, CH2, FROM, TO) {
-        while !self.is_done() {}
-
-        // Make sure that memory contents reflect what the user intended.
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // TODO: Use a tuple type?
-        (self.ch.0, self.ch.1, self.from, self.to)
-    }
-}
-
-impl<CH1, CH2, FROM, TO, WORD> DoubleBuffering<CH1, CH2, FROM, TO, ()>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD> + EndlessWriteTarget,
-{
-    /// Perform the next read of a double-buffered sequence
-    pub fn read_next<BUF: ReadTarget<ReceivedWord = WORD>>(
-        mut self,
-        buf: BUF,
-    ) -> DoubleBuffering<CH1, CH2, FROM, TO, ReadNext<BUF>> {
-        // Make sure that memory contents reflect what the user intended.
-        // TODO: How much of the following is necessary?
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Configure the _other_ DMA channel, but do not start it yet.
-        if self.second_ch {
-            self.ch.0.config(&buf, &mut self.to, self.pace, None, false);
-        } else {
-            self.ch.1.config(&buf, &mut self.to, self.pace, None, false);
-        }
-
-        // Chain the first channel to the second.
-        if self.second_ch {
-            self.ch.1.set_chain_to_enabled(&mut self.ch.0);
-        } else {
-            self.ch.0.set_chain_to_enabled(&mut self.ch.1);
-        }
-
-        DoubleBuffering {
-            ch: self.ch,
-            from: self.from,
-            to: self.to,
-            pace: self.pace,
-            state: ReadNext(buf),
-            second_ch: self.second_ch,
-        }
-    }
-}
-
-impl<CH1, CH2, FROM, TO, WORD> DoubleBuffering<CH1, CH2, FROM, TO, ()>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD> + EndlessReadTarget,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Perform the next write of a double-buffered sequence
-    pub fn write_next<BUF: WriteTarget<TransmittedWord = WORD>>(
-        mut self,
-        mut buf: BUF,
-    ) -> DoubleBuffering<CH1, CH2, FROM, TO, WriteNext<BUF>> {
-        // Make sure that memory contents reflect what the user intended.
-        // TODO: How much of the following is necessary?
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Configure the _other_ DMA channel, but do not start it yet.
-        if self.second_ch {
-            self.ch
-                .0
-                .config(&self.from, &mut buf, self.pace, None, false);
-        } else {
-            self.ch
-                .1
-                .config(&self.from, &mut buf, self.pace, None, false);
-        }
-
-        // Chain the first channel to the second.
-        if self.second_ch {
-            self.ch.1.set_chain_to_enabled(&mut self.ch.0);
-        } else {
-            self.ch.0.set_chain_to_enabled(&mut self.ch.1);
-        }
-
-        DoubleBuffering {
-            ch: self.ch,
-            from: self.from,
-            to: self.to,
-            pace: self.pace,
-            state: WriteNext(buf),
-            second_ch: self.second_ch,
-        }
-    }
-}
-
-impl<CH1, CH2, FROM, TO, NEXT, WORD> DoubleBuffering<CH1, CH2, FROM, TO, ReadNext<NEXT>>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD> + EndlessWriteTarget,
-    NEXT: ReadTarget<ReceivedWord = WORD>,
-{
-    /// Block until the the transfer is complete
-    pub fn wait(self) -> (FROM, DoubleBuffering<CH1, CH2, NEXT, TO, ()>) {
-        while !self.is_done() {}
-
-        // Make sure that memory contents reflect what the user intended.
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Invert second_ch as now the other channel is the "active" channel.
-        (
-            self.from,
-            DoubleBuffering {
-                ch: self.ch,
-                from: self.state.0,
-                to: self.to,
-                pace: self.pace,
-                state: (),
-                second_ch: !self.second_ch,
-            },
-        )
-    }
-}
-
-impl<CH1, CH2, FROM, TO, NEXT, WORD> DoubleBuffering<CH1, CH2, FROM, TO, WriteNext<NEXT>>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD> + EndlessReadTarget,
-    TO: WriteTarget<TransmittedWord = WORD>,
-    NEXT: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Block until transfer is complete
-    pub fn wait(self) -> (TO, DoubleBuffering<CH1, CH2, FROM, NEXT, ()>) {
-        while !self.is_done() {}
-
-        // Make sure that memory contents reflect what the user intended.
-        cortex_m::asm::dsb();
-        compiler_fence(Ordering::SeqCst);
-
-        // Invert second_ch as now the other channel is the "active" channel.
-        (
-            self.to,
-            DoubleBuffering {
-                ch: self.ch,
-                from: self.from,
-                to: self.state.0,
-                pace: self.pace,
-                state: (),
-                second_ch: !self.second_ch,
-            },
-        )
-    }
-}
-
-/// DMA configuration for sending and receiving data simultaneously
-pub struct BidirectionalConfig<CH1, CH2, FROM, BIDI, TO>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget,
-    BIDI: ReadTarget + WriteTarget,
-    TO: WriteTarget,
-{
-    ch: (CH1, CH2),
-    from: FROM,
-    bidi: BIDI,
-    to: TO,
-    from_pace: Pace,
-    to_pace: Pace,
-}
-
-impl<CH1, CH2, FROM, BIDI, TO, WORD> BidirectionalConfig<CH1, CH2, FROM, BIDI, TO>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    BIDI: ReadTarget<ReceivedWord = WORD> + WriteTarget<TransmittedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Create a DMA configuration for sending and receiving data simultaneously
-    pub fn new(
-        ch: (CH1, CH2),
-        from: FROM,
-        bidi: BIDI,
-        to: TO,
-    ) -> BidirectionalConfig<CH1, CH2, FROM, BIDI, TO> {
-        BidirectionalConfig {
-            ch,
-            from,
-            bidi,
-            to,
-            from_pace: Pace::PreferSink,
-            to_pace: Pace::PreferSink,
-        }
-    }
-
-    /// Set the transfer pacing for the DMA transfer from the source
-    pub fn from_pace(&mut self, pace: Pace) {
-        self.from_pace = pace;
-    }
-
-    /// Set the transfer pacing for the DMA transfer to the target
-    pub fn to_pace(&mut self, pace: Pace) {
-        self.to_pace = pace;
-    }
-
-    /// Start the DMA transfer
-    pub fn start(self) -> Bidirectional<CH1, CH2, FROM, BIDI, TO> {
-        // TODO
-        let _ch = self.ch;
-        let _from = self.from;
-        let _bidi = self.bidi;
-        let _to = self.to;
-        panic!("Not yet implemented.");
-    }
-}
-
-/// Instance of a bidirectional DMA transfer
-pub struct Bidirectional<CH1, CH2, FROM, BIDI, TO>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget,
-    BIDI: ReadTarget + WriteTarget,
-    TO: WriteTarget,
-{
-    ch: (CH1, CH2),
-    from: FROM,
-    bidi: BIDI,
-    to: TO,
-}
-
-impl<CH1, CH2, FROM, BIDI, TO, WORD> Bidirectional<CH1, CH2, FROM, BIDI, TO>
-where
-    CH1: SingleChannel,
-    CH2: SingleChannel,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    BIDI: ReadTarget<ReceivedWord = WORD> + WriteTarget<TransmittedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    /// Block until transfer is complete
-    pub fn wait(self) -> ((CH1, CH2), FROM, BIDI, TO) {
-        // TODO
-        // Reading self.value to satisfy clippy. Remove once implemented
-        let _ch = self.ch;
-        let _from = self.from;
-        let _bidi = self.bidi;
-        let _to = self.to;
-        panic!("Not yet implemented.");
-    }
-}
-
-/*impl<WORD, CH, FROM, TO> SingleBufferingConfig<CH, FROM, TO>
-where
-    CH: ChannelPair,
-    FROM: ReadTarget<ReceivedWord = WORD>,
-    TO: WriteTarget<TransmittedWord = WORD>,
-{
-    pub fn start_ring(mut self) -> Result<RingTransfer<CH, FROM, TO>, DMAError> {
-        // TODO: Fix the code for src_count/dest_count == 1.
-        let (src, src_count) = self.from.rx_address_count();
-        let src_incr = self.from.rx_increment();
-        let dest_incr = self.to.tx_increment();
-        let (dest, dest_count) = self.to.tx_address_count();
-        let transfer_size = mem::size_of::<WORD>() as u32;
-
-        // Buffers with increment need to be aligned.
-        if src_incr && !Self::check_alignment(src, src_count * transfer_size) {
-            return Err(DMAError::Alignment);
-        }
-        if dest_incr && !Self::check_alignment(dest, dest_count * transfer_size) {
-            return Err(DMAError::Alignment);
-        }
-        // Memory-to-memory copies are not supported.
-        if src_incr && dest_incr {
-            return Err(DMAError::IllegalConfig);
-        }
-        // Additional sanity checks - we have to completely transfer all memory buffers.
-        if src_incr && src_count > dest_count {
-            return Err(DMAError::IllegalConfig);
-        }
-        if dest_incr && dest_count > src_count {
-            return Err(DMAError::IllegalConfig);
-        }
-
-        const TREQ_UNPACED: u8 = 0x3f;
-        let treq = match self.pace {
-            Pace::PreferSource => FROM::rx_treq().or(TO::tx_treq()).unwrap_or(TREQ_UNPACED),
-            Pace::PreferSink => TO::tx_treq().or(FROM::rx_treq()).unwrap_or(TREQ_UNPACED),
-        };
-
-        // We split buffers in two halves for the two DMA channels.
-        let (src1, src2) = if src_incr {
-            (src, src + src_count * transfer_size)
-        } else {
-            (src, src)
-        };
-        let (dest1, dest2, wrap_dest) = if src_incr {
-            (dest, dest + dest_count * transfer_size, true)
-        } else {
-            (dest, dest, false)
-        };
-        let len = u32::min(src_count, dest_count);
-        let ring_size = if src_incr || dest_incr {
-            (len * transfer_size / 2).trailing_zeros() as u8
-        } else {
-            0
-        };
-        // Configure the second DMA channel - we do not write any trigger register as we do not
-        // want to immediately start the transfer.
-        self.ch
-            .ch2()
-            .ch_read_addr
-            .write(|w| unsafe { w.bits(src2) });
-        self.ch
-            .ch2()
-            .ch_trans_count
-            .write(|w| unsafe { w.bits(len / 2) });
-        self.ch
-            .ch2()
-            .ch_write_addr
-            .write(|w| unsafe { w.bits(dest2) });
-        self.ch.ch2().ch_al1_ctrl.write(|w| unsafe {
-            w.data_size()
-                .bits(mem::size_of::<WORD>() as u8 >> 1)
-                .incr_read()
-                .bit(src_incr)
-                .incr_write()
-                .bit(dest_incr)
-                .treq_sel()
-                .bits(treq)
-                .en()
-                .bit(true)
-                .chain_to()
-                .bits(self.ch.id())
-                .ring_sel()
-                .bit(wrap_dest)
-                .ring_size()
-                .bits(ring_size)
-        });
-        // Configure the first DMA channel - the access to CTRL_TRIG starts the transfer.
-        self.ch.ch().ch_read_addr.write(|w| unsafe { w.bits(src1) });
-        self.ch
-            .ch()
-            .ch_trans_count
-            .write(|w| unsafe { w.bits(len / 2) });
-        self.ch
-            .ch()
-            .ch_write_addr
-            .write(|w| unsafe { w.bits(dest1) });
-        self.ch.ch().ch_ctrl_trig.write(|w| unsafe {
-            w.data_size()
-                .bits(mem::size_of::<WORD>() as u8 >> 1)
-                .incr_read()
-                .bit(src_incr)
-                .incr_write()
-                .bit(dest_incr)
-                .treq_sel()
-                .bits(treq)
-                .en()
-                .bit(true)
-                .chain_to()
-                .bits(self.ch.id2())
-                .ring_sel()
-                .bit(wrap_dest)
-                .ring_size()
-                .bits(ring_size)
-        });
-
-        Ok(RingTransfer {
-            ch: self.ch,
-            from: self.from,
-            to: self.to,
-        })
-    }
-
-    fn check_alignment(addr: u32, len: u32) -> bool {
-        if !len.is_power_of_two() {
-            return false;
-        }
-        if (addr & (len - 1)) != 0 {
-            return false;
-        }
-        true
-    }
-}
-
-pub struct RingTransfer<CH: ChannelPair, FROM: ReadTarget, TO: WriteTarget> {
-    ch: CH,
-    from: FROM,
-    to: TO,
-}
-
-impl<CH, FROM, TO> RingTransfer<CH, FROM, TO>
-where
-    CH: ChannelPair,
-    FROM: ReadTarget,
-    TO: WriteTarget,
-{
-    // TODO
 }*/
 
 /// Pacing for DMA transfers.
